Establish a game_id-based API on GameMaster mutating game state
What I had previously was a temporary sketch of an interface; this
change to GameMaster.advance/3 is more appropriate, given our
intended architecture.

Clients of GameMaster will advance a game only by providing a game_id
(a UUID binary), the player index, and a key_code.

What happens inside the GameMaster, the single source of truth, is
a black box, as far as the client (e.g. a LiveView process) is
concerned.